+++ api/src/main/java/net/jqwik/api/arbitraries/MapArbitrary.java
@@ -5,6 +5,8 @@ import java.util.function.*;
 
 import org.apiguardian.api.*;
 
+import net.jqwik.api.*;
+
 import static org.apiguardian.api.API.Status.*;
 
 /**
@@ -12,7 +14,7 @@ import static org.apiguardian.api.API.Status.*;
  * of type {@linkplain Map}
  */
 @API(status = MAINTAINED, since = "1.3.2")
-public interface MapArbitrary<K, V> extends SizableArbitrary<Map<K, V>> {
+public interface MapArbitrary<K, V> extends Arbitrary<Map<K, V>>, SizableArbitrary<Map<K, V>> {
 
 	/**
 	 * Fix the size to {@code size}.